クラス名変更、メソッドの移動など
